fixes according comments
