+++ src/test/java/utils/SqlDataProvider.java
@@ -1,6 +1,6 @@
 package utils;
 
-public class SQLDataProvider {
+public class SqlDataProvider {
     public static final String DBNAME = "tu06";
     public static final String DBPASSWORD = "6Qazwsx!";
     public static final String URL = "jdbc:mysql://test.biz.ua:3306/orangehrm_mysql";
